fix(cua-driver): Linux X11 and Wayland convergence (#2182)
* fix(cua-driver): Linux X11 and Wayland convergence (conv2 4-5/7)

* fix(cua-driver): correct Wayland targeting and chords

* fix(cua-driver): preserve Wayland physical hotkeys

* fix(cua-driver): wire generated Wayland protocol

* fix(cua-driver): translate Wayland web content insets

* fix(cua-driver): align Sway web input coordinates

* fix(cua-driver): await WebKit origin fallback outside closure

* fix(cua-driver): unify Sway web coordinate geometry

* fix(cua-driver): reconcile Sway window and content origins

* test(cua-driver): refresh live Wayland fixture geometry

* fix(cua-driver): identify roleless WebKit subtrees

+++ libs/cua-driver/rust/crates/platform-linux/src/wayland/persistent_vptr.rs
@@ -147,7 +147,7 @@ fn handle_press(
     // Keep the (out_w, out_h) but drop the queue + state at end of scope; the
     // vptr itself remains alive (Wayland objects survive their original queue
     // as long as the Connection is alive).
-    let mut sess = open_vptr_session(Some(window_id as u32))?;
+    let mut sess = open_vptr_session(Some(window_id))?;
     let (w, h) = (sess.output_w, sess.output_h);
     let px = x.clamp(0, w as i32 - 1) as u32;
     let py = y.clamp(0, h as i32 - 1) as u32;
